refactor(update): consolidate into single cli-update module
